Add async Kafka payment system, WebSocket notifications, and comprehensive CI test suite

+++ server/src/main/java/com/example/bank/service/AccountService.java
@@ -1,9 +1,10 @@
 package com.example.bank.service;
 
-import com.example.bank.dto.AccountCreateDto;
-import com.example.bank.dto.AccountDto;
+import com.example.bank.dto.Account.AccountCreateDto;
+import com.example.bank.dto.Account.AccountDto;
 
 import java.util.List;
+import java.math.BigDecimal;
 
 public interface AccountService {
 
@@ -11,15 +12,15 @@ public interface AccountService {
 
     AccountDto getAccountById(Long id);
 
-    AccountDto deposit(Long id, double amount);
+    AccountDto deposit(Long id, BigDecimal amount);
 
-    AccountDto withdraw(Long id, double amount);
+    AccountDto withdraw(Long id, BigDecimal amount);
 
     List<AccountDto> getAllAccounts();
 
     void deleteAccount(Long id);
 
-    void transfer(Long fromId, Long toId, double amount);
+    void transfer(Long fromId, Long toId, BigDecimal amount);
 
     boolean isAccountOwner(String username, Long accountId);
 }
